+++ src/calendar.js
@@ -1707,6 +1707,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
         event.id = _element_ID_Event_Widget_Day + eventDetails.id;
         event.setAttribute( "event-type", getNumber( eventDetails.type ) );
         event.setAttribute( "event-id", eventDetails.id );
+        event.setAttribute( "event-is-all-day", eventDetails.isAllDay );
 
         if ( !fireCustomTrigger( "onWidgetEventRender", event, eventDetails ) ) {
             var title = createElement( "div", "title" ),
@@ -2513,6 +2514,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             event.ondblclick = cancelBubble;
             event.setAttribute( "event-type", getNumber( eventDetails.type ) );
             event.setAttribute( "event-id", eventDetails.id );
+            event.setAttribute( "event-is-all-day", eventDetails.isAllDay );
 
             event.onclick = function ( e ) {
                 increaseEventZIndex( e, event );
@@ -3127,6 +3129,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
 
             event.setAttribute( "event-type", getNumber( eventDetails.type ) );
             event.setAttribute( "event-id", eventDetails.id );
+            event.setAttribute( "event-is-all-day", eventDetails.isAllDay );
             column.appendChild( event );
 
             event.onclick = function ( e ) {
@@ -4307,6 +4310,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
     
                         event.setAttribute( "event-type", getNumber( eventDetails.type ) );
                         event.setAttribute( "event-id", eventDetails.id );
+                        event.setAttribute( "event-is-all-day", eventDetails.isAllDay );
 
                         if ( !_options.useOnlyDotEventsForMainDisplay ) {
                             if ( !fireCustomTrigger( "onFullMonthEventRender", event, eventDetails ) ) {
@@ -4898,6 +4902,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
             event.id = _element_ID_Event_Month + eventDetails.id;
             event.setAttribute( "event-type", getNumber( eventDetails.type ) );
             event.setAttribute( "event-id", eventDetails.id );
+            event.setAttribute( "event-is-all-day", eventDetails.isAllDay );
 
             if ( !fireCustomTrigger( "onAllEventsEventRender", event, eventDetails ) ) {
                 var title = createElement( "div", "title" ),
@@ -5361,6 +5366,7 @@ function calendarJs( elementOrId, options, searchOptions ) {
         event.id = _element_ID_Event_Timeline_Day + eventDetails.id;
         event.setAttribute( "event-type", getNumber( eventDetails.type ) );
         event.setAttribute( "event-id", eventDetails.id );
+        event.setAttribute( "event-is-all-day", eventDetails.isAllDay );
         timelineRowItems.appendChild( event );
 
         if ( !fireCustomTrigger( "onTimelineEventRender", event, eventDetails ) ) {